Extract P2P coordinator worker tests
